Moves tests to their own directory. Adds codeclimate badge.
